sync with msgilligan/graalvm-oracle-25ea
msgilligan/graalvm-oracle-25ea was rebased and force-pushed.
I don't want to do that on the `jdk-ea-25` branch so this commit
manually synchronizes with the updated msgilligan/graalvm-oracle-25ea.

There's probably a better way of doing this using git, but I can't think
of one at the moment, so this will do for now.